requirements.txt
ray==2.3.1
torch==2.0.0
transformers==4.27.1
click
ipython
ipywidgets
matplotlib
numpy
pandas
PyYAML
seaborn
tabulate
tqdm
thop
